Implement chunking to support smaller LUTs.
This works by breaking each field into N chunks during encoding, allowing a LUT sized to the chunk to be used, rather than to the width of the entire field.

Since the LUT operates on chunks, the more bits in each chunk, the larger the LUT (size is  2^#bits * entry size). The consequence of using smaller chunks is that more bitwise operations are required during an encode (since chunks must be isolated for lookup, and their results aggregated).

In addition to #bits, #chunks must now be specified. The #bits now designates the number of bits in each chunk. For example, to handle 16 bit fields, you could specify:

    - 1 chunk, 16 bits: biggest LUT, fastest
    - 2 chunks, 8 bits:
    - 4 chunks, 4 bits: smallest LUT, slowest

Template expansion was used to implement chunked encoding, so you will incur only exactly the number of operations required for the #chunks you specify. For example, specifying #chunks = 1 is equivalent to disabling chunking completely.