Remove restriction on LUT table size.
Chunk count is no longer user-specified. Instead, specify the number of bits in a field.

"Bits" has been renamed to "LutBits" but serves the same purpose (it's the size of a chunk, the unit of lookup). The number of chunks used is calculated based on LutBits and the field size (FieldBits). These changes allow any LUT size to be used, even if the field width (FieldBits) is not a multiple of the chunk size (LutBits).

Note that the upper bits (above FieldBits) of an input field must now be cleared manually prior to encoding (if not already). A mask exists as a static member of MortonNDEncoder named "InputMask" to make this easy: simply '&' each input field with 'MortonNDEncoder<...>::InputMask' when calling the Encode method.