MortonNDLutEncoder: InputMask is now a constexpr function.
This change is required to use the encoder with non class-literal user-defined classes, since it's not possible to store these in a constexpr class.

This change also ensures that InputMash() is not available for user-defined class types, since the current method of creation relies on an unsigned integer type of a known width.