MortonNDLutEncoder: Require std::size_t be only constructible from T, instead of convertible to.
Originally, we required T to be implicitly convertible to std::size_t, which doesn't work well in practice.

The biggest issue is that if a user-defined class does not provide the numeric operators we need, it will be implicitly converted to std::size_t (and no error will be generated), which we certainly don't want. I was also unable to find BigInteger-like implementation which supplied implicit conversion operators.

By requiring instead that std::size_t is constructible from T, it's sufficient for T to define an explicit conversion operator. We just need to add std::size_t(...) around any expression resulting in a T where we want an std::size_t.